+++ src/hook/useInfiniteScroll.tsx
@@ -1,0 +1,44 @@
+"use client";
+import { useCallback, useEffect, useRef } from "react";
+
+interface UseInfiniteScrollProps {
+  fetchNextPage: () => void;
+  hasNextPage: boolean;
+  isLoading: boolean;
+}
+
+const useInfiniteScroll = ({
+  fetchNextPage,
+  hasNextPage,
+  isLoading
+}: UseInfiniteScrollProps) => {
+  const loadMoreRef = useRef<HTMLDivElement>(null);
+
+  const handleObserver = useCallback(
+    (entries: IntersectionObserverEntry[]) => {
+      const target = entries[0];
+      if (target.isIntersecting && hasNextPage && !isLoading) {
+        fetchNextPage();
+      }
+    },
+    [fetchNextPage, hasNextPage, isLoading]
+  );
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(handleObserver, {
+      root: null,
+      rootMargin: "20px",
+      threshold: 1.0
+    });
+
+    if (loadMoreRef.current) observer.observe(loadMoreRef.current);
+
+    return () => {
+      if (loadMoreRef.current) observer.unobserve(loadMoreRef.current);
+    };
+  }, [handleObserver]);
+
+  return loadMoreRef;
+};
+
+export default useInfiniteScroll;
